requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dateparser==1.2.0
dj-database-url==3.0.1
Django==5.1.3
django-cors-headers==4.6.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
multidict==6.1.0
phonenumbers==9.0.12
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyotp==2.9.0
pypng==0.20220715.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qrcode==7.4.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
service-identity==24.2.0
setuptools==75.4.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websockets==14.1
whitenoise==6.10.0
yarl==1.17.1
zope.interface==7.1.1
